Updated the Exploratory Analysis Section
I added more distributions in the "Variable Analysis" section and added more boxplots in the "Categorical Variable Analysis" section. I also modified the format to make it more legible in the markdown. There are 2 reasons why I added more graphs in our exploratory analysis. 1) It gives the reader a chance to observe all of the variables that we included in our first iteration of the model. 2) In the instructions, Eric stated that he wants "an analysis of the relationship between the levels of each categorical variable and the outcome variable". That said, I included "Total Paid by Employment Length" boxplot and "Total Paid by Loan Status".