Added the Exploratory Analysis Work in Script
Added the Exploratory Analysis work in the script. There are plots that are not present in the markdown. I commented on those plots "Not included in the markdown". The reason I didn't include all plots is because the output was insufficient and wouldn't mean much to the reader. I could be wrong, so feel free to check them out and put them in the markdown if it makes sense to you.